tests/weak_semaphore_tests.rs
#[cfg(test)]
mod weak_semaphore_tests {
    use std::sync::{Arc, Mutex};
    use std::thread;
    use std::time::{Duration};
    use concurrency_project::semaphore::weak_semaphore::WeakSemaphore;

    // 동시성 제어 테스트
    #[test]
    fn test_weak_semaphore_concurrency() {
        let semaphore = Arc::new(WeakSemaphore::new(1));
        let counter = Arc::new(Mutex::new(0));
        let num_threads = 10; // 현실적인 스레드 수로 조정
        let mut handles = vec![];

        for _ in 0..num_threads {
            let sem = Arc::clone(&semaphore);
            let counter = Arc::clone(&counter);
            handles.push(thread::spawn(move || {
                sem.acquire();
                {
                    let mut count = counter.lock().unwrap();
                    *count += 1;
                    assert_eq!(*count, 1, "More than one thread in critical section");
                    thread::sleep(Duration::from_millis(10));
                    *count -= 1;
                }
                sem.release();
            }));
        }

        for handle in handles {
            handle.join().unwrap();
        }

        assert_eq!(*counter.lock().unwrap(), 0, "Final count should be 0");
    }

    // 공정성 테스트
    #[test]
    fn test_weak_semaphore_fairness() {
        let semaphore = Arc::new(WeakSemaphore::new(1));
        let order = Arc::new(Mutex::new(Vec::new()));
        let num_threads = 5;
        let mut handles = vec![];

        for i in 0..num_threads {
            let sem = Arc::clone(&semaphore);
            let order = Arc::clone(&order);
            handles.push(thread::spawn(move || {
                sem.acquire();
                order.lock().unwrap().push(i);
                thread::sleep(Duration::from_millis(10));
                sem.release();
            }));
            thread::sleep(Duration::from_millis(1)); // 스레드 생성 간 지연 추가
        }

        for handle in handles {
            handle.join().unwrap();
        }

        let final_order = order.lock().unwrap();
        println!("Execution order: {:?}", *final_order);
        // Weak semaphore는 특정 순서를 보장하지 않음
        assert_eq!(final_order.len(), num_threads, "All threads should have executed");
    }

    // 초기화 상태 테스트
    #[test]
    fn test_weak_semaphore_zero_init() {
        let semaphore = Arc::new(WeakSemaphore::new(0));
        let flag = Arc::new(Mutex::new(false));
        let handle = {
            let sem = Arc::clone(&semaphore);
            let flag = Arc::clone(&flag);
            thread::spawn(move || {
                sem.acquire();
                *flag.lock().unwrap() = true;
                sem.release();
            })
        };

        thread::sleep(Duration::from_millis(100));
        assert_eq!(*flag.lock().unwrap(), false, "Semaphore should not have been acquired");

        semaphore.release();
        handle.join().unwrap();
        assert_eq!(*flag.lock().unwrap(), true, "Semaphore should have been acquired and released");
    }
}

// 동시성 제어: test_weak_semaphore_concurrency는 한 번에 하나의 스레드만 임계 영역에 접근할 수 있음을 확인합니다.
// 공정성: test_weak_semaphore_fairness는 스레드 실행 순서가 약한 세마포어의 특성상 비결정적일 수 있음을 보여줍니다. 모든 스레드가 실행되었는지를 체크합니다.
// 초기화 상태: test_weak_semaphore_zero_init는 세마포어가 0으로 초기화되었을 때의 동작을 확인합니다. 세마포어가 해제되기 전까지 스레드가 블록되어야 합니다.
// 타이밍 검증: test_weak_semaphore_timing은 각 스레드가 세마포어를 획득하고 해제하는 시간을 기록하여 동작을 시각화합니다.